refactor: 패키지 분리 및 불필요한 getter, setter 제거
- 웹과 콘솔 패키지로 구분하여 클래스 분류
