--- src/main/java/racingcar/Application.java
@@ -1,10 +1,0 @@
-package racingcar;
-
-import racingcar.controller.Controller;
-
-public class Application {
-    public static void main(String[] args) {
-        Controller controller = new Controller();
-        controller.runGame();
-    }
-}
