+++ src/main/java/racingcar/console/ConsoleApplication.java
@@ -1,0 +1,10 @@
+package racingcar.console;
+
+import racingcar.console.controller.ConsoleController;
+
+public class ConsoleApplication {
+    public static void main(String[] args) {
+        ConsoleController consoleController = new ConsoleController();
+        consoleController.runGame();
+    }
+}
