+++ src/main/java/racingcar/console/view/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package racingcar.view;
+package racingcar.console.view;
 
 public enum ExceptionMessage {
     CAR_NAME_NOT_INPUT_MESSAGE("자동차 이름을 입력해주세요."),
