+++ src/main/java/racingcar/console/view/InputView.java
@@ -1,4 +1,4 @@
-package racingcar.view;
+package racingcar.console.view;
 
 import java.util.Scanner;
