+++ src/main/java/racingcar/web/WebApplication.java
@@ -1,12 +1,11 @@
-package racingcar;
+package racingcar.web;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class RacingCarApplication {
+public class WebApplication {
     public static void main(String[] args) {
-        SpringApplication.run(RacingCarApplication.class, args);
+        SpringApplication.run(WebApplication.class, args);
     }
-
 }
