+++ src/main/java/racingcar/web/controller/dto/GameResultDto.java
@@ -1,4 +1,4 @@
-package racingcar.controller.dto;
+package racingcar.web.controller.dto;
 
 import java.util.List;
