+++ src/main/java/racingcar/web/util/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.web.util;
 
 public interface NumberGenerator {
     int generateNumber();
