+++ src/main/java/racingcar/web/util/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.web.util;
 
 import org.springframework.stereotype.Component;
