+++ src/main/java/racingcar/web/util/TestNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.web.util;
 
 import java.util.List;
