+++ src/test/java/racingcar/dao/ResultDaoTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import racingcar.web.dao.ResultDao;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
