+++ src/test/java/racingcar/domain/CarTest.java
@@ -2,6 +2,8 @@ package racingcar.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.web.domain.Car;
+import racingcar.web.domain.Name;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
